disable CI run for 1.9.2
strange failure that I can't seem to reproduce locally on the same version of
Ruby. I'm not sure of the cause but given that it is isolated to 1.9.2, and
cannot be reproduced on my machine I don't feel like it is a problem in xpool
itself. I studied the code around the exception and I cannot see any fault
that could be placed on xpool.

Travis CI failed build:
https://travis-ci.org/robgleeson/xpool/builds/5753926